api/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-CORS==4.0.0

# Database
PyMySQL==1.1.0
cryptography==41.0.7

# Timezone handling
pytz==2023.3

# Image processing (for QR codes)
Pillow==10.1.0
qrcode==7.4.2

# Password hashing
bcrypt==4.0.1

# HTTP requests for external APIs
requests==2.31.0

# Optional: Rate limiting (uncomment if needed)
# Flask-Limiter==3.5.0

# Development dependencies (not needed for production)
# python-dotenv==1.0.0
Werkzeug==2.3.7

# Production server
gunicorn==21.2.0

# Email dependencies
blinker==1.6.2

# Additional dependencies for Vercel
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
Jinja2==3.1.2
